Reformatted part 1 code to instead use the push method for arrays.
I wanted to practice the push function to understand it, so I reformatted my code for part one to simply push strings into the array. This got rid of a lot of variables and conditions, as the push method actually declared the string before I can write into the array.  This avoids a lot of undefined entries in the array and determines the length of the array.